tools/ingest_notebooklm/requirements.txt
notebooklm
requests
firebase-admin
pyyaml
